% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl.R
\name{bl}
\alias{bl}
\title{Baseline Correction}
\usage{
bl(X, lambda = 1e+07, it_max = 30)
}
\arguments{
\item{X}{The X matrix you wish to correct the baseline of}

\item{lambda}{The smoothing parameter}

\item{it_max}{The maximum iterations before bl stops}
}
\value{
A matrix containing spectra with corrected baselines
}
\description{
Baseline Correction
}
\examples{
path = system.file('extdata', package = 'unorm')
X_bl <- bl(X)
}
\seealso{
Other {Data Minipulation}: 
\code{\link{binin}()},
\code{\link{flip}()},
\code{\link{preprocessing}()}
}
\concept{{Data Minipulation}}
