% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creNorm.R
\name{creNorm}
\alias{creNorm}
\title{Creatinine Normalisation}
\usage{
creNorm(X, ppm, cre3 = c(3.043, 3.055), cre4 = c(4.05, 4.07))
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{ppm}{A numerical array holding the chemical shift values of the X matrix}

\item{cre3}{A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 3.05 starts and ends}

\item{cre4}{A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 4.05 starts and ends}
}
\value{
A list of the normalised X matrix, a numerical array of the corresponding dilution factors, and ratio of the creatinine signals (ideally should equal 0.667)
}
\description{
Creatinine Normalisation (CN) is a useful method much like region of interest normalisation that can normalise spectra based on the total area of the creatinine signal at the chemical shift 3.05ppm.
}
\details{
#Limitations:
\itemize{
\item While useful, there are complicating factors associated with CN's use such as inter- and intra-individual creatinine signal fluctuation which hinders accurate normalisation.
\item CN relies on the idea that creatinine concentrations do not change in urine samples, the only changes to creatinine signal are created by effects of dilution but this is incorrect.
\item Creatinine concentrations fluctuate depending on a person's physiology and behaviour so use CN consciously.
}
}
\examples{
\dontrun{
  Xn <- creNorm(X)
  creDilf <- Xn[[2]]
  Xn <- Xn[[1]]}
}
\seealso{
Other {Attribute-Based}: 
\code{\link{roiNorm}()},
\code{\link{taNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
