% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmNorm.R
\name{hmNorm}
\alias{hmNorm}
\title{Histogram Matching}
\usage{
hmNorm(
  X,
  ppm,
  noi = c(10, 11),
  intensity_binwidth = 0.1,
  alpha_from = 0.5,
  alpha_to = 1.5,
  alpha_n = 101,
  use_median = F
)
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{ppm}{A numerical array holding the chemical shift values of the X matrix}
}
\value{
A list with the normalised X matrix and an array of the corresponding dilution factors.
}
\description{
\code{\link[=hmNorm]{hmNorm()}} is a reference based normalisation that aims to match the histogram created from a experimental spectra with that made from a reference spectra.
}
\details{
#Limitations:
\itemize{
\item Histogram matching (HM) overcomes issues like peak shift that impact methods such as PQN, however, HM has limitations caused by differently shaped histograms.
\item Because not all histograms will have the same level of skewness or kurtosis, not all matchings will be ideal which limits this methods applicability.
}
}
\examples{
\dontrun{
 histMod <- hmNorm(X, ppm)
 hmDilf <- histMod[[2]]
 Xn <- histMod[[1]]}
}
\seealso{
The methods paper for HM can be found here: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}

Other {Reference-Based}: 
\code{\link{pqNorm}()},
\code{\link{xfNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
