% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqn.R
\name{pqn}
\alias{pqn}
\title{Probabilistic Quotient Normalisation}
\usage{
pqn(X)
}
\arguments{
\item{X}{spectral matrix with rows being the spectra and columns being the chemical shift variables}
}
\value{
list of normalised X matrix and an array of the corresponding dilution factors
}
\description{
Probabilistic Quotient Normalisation
}
\details{
https://doi.org/10.1021/ac051632c
}
\author{
kylebario1@gmail.com
}
