% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqNorm.R
\name{pqNorm}
\alias{pqNorm}
\title{Probabilistic Quotient Normalisation (PQN)}
\usage{
pqNorm(X)
}
\arguments{
\item{X}{The numerical matrix containing the NMR data. The rows contain information of one whole spectrum and the columns contain the specific chemical shift variables.}
}
\value{
The output of this function is a list containing the normalised version of X and a numerical array of the corresponding dilution factors calculated by the function.
}
\description{
PQN is currently the gold standard normalisation method use to transform NMR spectra.
}
\details{
\subsection{How It Works:}{
\itemize{
\item PQN works by normalising experimental spectra in the provided X matrix in relation to a reference spectrum.
\item \code{\link[=pqNorm]{pqNorm()}} creates the reference automatically by calculating the median spectrum of X as outlined in the initial methods paper (see 'See also')
\item \code{\link[=pqNorm]{pqNorm()}} derives a quotient for each ppm value in a experimental spectrum by dividing it's intensities with that of the reference spectrum's.
\item The most frequently occurring quotient is calculated and is said to be the dilution coefficient (\code{dilf}) of that spectrum.
\item The sample is then scaled with this dilf and will be comparable with all other spectra normalised with the reference spectrum.
}
}

\subsection{Background Information:}{
\itemize{
\item PQN is currently the gold standard for normalising NMR spectra.
\item Multiple studies have tested its validity.
\item PQN reliably normalises spectra and handles large amounts of noise.
}
}

\subsection{Limitations:}{
\itemize{
\item PQN's biggest limitation is that it operates on the assumption that the majority of the spectra will stay constant so hugely varying spectra may be suboptimally normalised.
}
}
}
\examples{
\donttest{
Xn <- pqNorm(X)
pqnDilf <- Xn[[2]]
Xn <- t(Xn[[1]])
}
}
\seealso{
The methods paper first describing PQN can be found here: \url{https://doi.org/10.1021/ac051632c}

Other {Reference-Based}: 
\code{\link{hmNorm}()},
\code{\link{xfNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
