% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qNorm.R
\name{qNorm}
\alias{qNorm}
\title{Quantile Normalisation}
\usage{
qNorm(X)
}
\arguments{
\item{X}{A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables}
}
\value{
This function returns a list with:
\enumerate{
\item The normalised X matrix based on the calculated dilf in the first element,
\item The quantile normalised X matrix in the second element, and
\item The dilf calculated in the third element.
}
}
\description{
A method of NMR spectral normalisation where the maximum intensities
}
\details{
The intensities of each spectrum are ordered smallest to largest, the means of these 'quantiles' are calculated and these means are then reassigned to the ppm they match to (i.e., the ppm which contained the highest intensity will be assigned the mean of the highest values and so on)
}
\examples{
data(X)
qNorm(X)
}
\seealso{
The methods paper that describes quantile normalisation: \url{https://doi.org/10.1007/s11306-011-0350-z}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{qNorm2}()},
\code{\link{roiNorm}()},
\code{\link{taNorm}()},
\code{\link{vecNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
