% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qNorm2.R
\name{qNorm2}
\alias{qNorm2}
\title{Quantile Normalisation 2}
\usage{
qNorm2(X, uv_used = "mode")
}
\arguments{
\item{X}{A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables}
}
\value{
This function returns a list with:
\enumerate{
\item The normalised X matrix based on the calculated dilf in the first element,
\item The quantile normalised X matrix in the second element, and
\item The dilf calculated in the third element.
}
}
\description{
A method of NMR spectral normalisation where the maximum intensities
}
\details{
\subsection{How It Works:}{
}

\subsection{Advantages:}{
}

\subsection{Limitations:}{
}
}
\examples{
Xqn <- qNorm(X)
Xn <- Xqn$Xn
Xq <- Xqn$Xq
qDilf <- Xqn$dilf
}
\seealso{
The methods paper that describes quantile normalisation: \url{https://doi.org/10.1007/s11306-011-0350-z}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{qNorm}()},
\code{\link{roiNorm}()},
\code{\link{taNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
