% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roiNorm.R
\name{roiNorm}
\alias{roiNorm}
\title{Region of Interest Normalisation (ROI)}
\usage{
roiNorm(X, ppm, shift = c(3, 3.1))
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{ppm}{A numerical array holding the chemical shift values of the X matrix}

\item{shift}{The numerical values defining the lower and upper regions of the Region of Interest. default = c(3,3.1).}
}
\value{
this function returns a list of the normalised X matrix and a numerical array of the corresponding dilution factors.
}
\description{
This function, \code{\link[=roiNorm]{roiNorm()}}, normalises the spectra based on a specific area of the spectra. This is helpful when one region is constant across all spectra.
}
\details{
http://dx.doi.org/10.1007/s11306-018-1400-6
}
\examples{
\dontrun{
  Xn <- roiNorm(X, ppm, shift = c(2,2.3))
  roiDilf <- Xn[[2]]
  Xn <- Xn[[1]]}
}
\seealso{
Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{taNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
