% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taNorm.R
\name{taNorm}
\alias{taNorm}
\title{Total Area Normalisation}
\usage{
taNorm(X)
}
\arguments{
\item{X}{A matrix containing the NMR spectra with rows contain all data points from one spectra and the columns containing the chemical shift variables.}
}
\value{
A list with the normalised X matrix and an array of the corresponding dilution factors.
}
\description{
Total Area Normalisation
}
\details{
also called Constant Sum or Integral normalisation
}
\author{
kylebario1@gmail.com
}
