% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecNorm.R
\name{vecNorm}
\alias{vecNorm}
\title{Vector Length Normalisation}
\usage{
vecNorm(X)
}
\arguments{
\item{X}{The spectr(a/um) to be normalised}
}
\value{
The return value of \code{vecNorm()} will be of same shape as the X argument parsed to it, either an array or matrix of normalised X
}
\description{
This function aims to normalise NMR spectra by their vector length.
}
\details{
This function operates by squaring all values of the spectrum, summing them, and square rooting the result. Then each element from that spectrum is divided by this vector length.
}
\examples{
data(X)
vecNorm(X)
cat(dilf_vec)
}
\seealso{
The methods paper that describes this method of normalisation can be found here: \url{https://doi.org/10.1021/ac051632c}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{qNorm}()},
\code{\link{roiNorm}()},
\code{\link{taNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
