% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfNorm.R
\name{xfNorm}
\alias{xfNorm}
\title{Normalisation based on an External Factor}
\usage{
xfNorm(X, xfactor)
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{xfactor}{A numerical array of values that correspond to the dilution coefficients of the spectra. Must be row-matched to the provided X matrix.}
}
\value{
This function returns a list of the normalised X matrix and a numerical array of the corresponding dilution factors
}
\description{
This function utilises an array of dilution coefficients acquired externally to normalise spectra.
}
\section{How It Works:}{
All the intensities of a spectrum are divided by its external dilution coefficient.
\subsection{Advantages:}{
\itemize{
\item Reliable: dilution estimations from external sources such as freezing-point depression osmometry are highly accurate, reducing the error in normalisation.
Normalisation based on an external factor is a more robust method of normalisation for urine spectra especially using the gold standard of dilution measurement - freezing-point depression osmometry - but this requires more effort and time compared to other normalisation methods like PQN.
}
}
}

\examples{
\dontrun{
  Xn <- xfNorm(X, osmo)
  xfDilf <- Xn[[2]]
  Xn <- Xn[[1]]}
}
\seealso{
\url{https://doi.org/10.1021/ac051632c}

Other {Reference-Based}: 
\code{\link{hmNorm}()},
\code{\link{pqNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
