% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creNorm.R
\name{creNorm}
\alias{creNorm}
\title{Creatinine Normalisation}
\usage{
creNorm(X, ppm, cre3 = c(3.043, 3.055), cre4 = c(4.05, 4.07))
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{ppm}{A numerical array holding the chemical shift values of the X matrix}

\item{cre3}{A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 3.05 starts and ends}

\item{cre4}{A concatenated numerical value of the lower and upper ppm values where the creatinine peak at 4.05 starts and ends}
}
\value{
A list of the normalised X matrix, a numerical array of the corresponding dilution factors, and ratio of the creatinine signals (ideally should equal 0.667)
}
\description{
Creatinine Normalisation
}
\details{
Creatinine normalisation is a method used to normalise based on the intensities of the creatinine signal. While useful, there are complicating factors associated with its use such as inter- and intra-individual creatinine signal fluctuation which hinders accurate normalisation.
}
\examples{
  Xn <- creNorm(X)
  creDilf <- Xn[[2]]
  Xn <- Xn[[1]]
}
\author{
\email{kylebario1@gmail.com}
}
