% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmNorm.R
\name{hmNorm}
\alias{hmNorm}
\title{Histogram Matching}
\usage{
hmNorm(
  X,
  noi,
  int_binwid = 0.1,
  alpha = c(0.1, 2),
  use_median = NULL,
  tol = 1e-05
)
}
\arguments{
\item{X}{A numerical matrix containing the NMR spectra to be normalised. Rows should be the spectra and columns being the chemical shift variables}

\item{noi}{The array of maximum noise estimations produced from the function \code{\link[=noise]{noise()}} or provided by the user. Must match X rows.}

\item{int_binwid}{This argument dictates the width of the bins. The average span of intensities is from \code{10}-\code{30}, meaning that an \code{intensity_binwidth} of \code{0.1} would give you 200 bins.}

\item{alpha}{The lower and upper bounds that the golden selection search will search between}

\item{use_median}{This argument dictates whether the function will calculate the median and use that as the reference spectrum or not. If set to \code{FALSE}, the first sample will be used as the reference. This practice is outlined in the methods paper (see 'See also')}

\item{tol}{This defines the tolerance or level of precision the golden selection search will search until. (i.e., it will search until the bounds are \code{tol} distance apart)}
}
\value{
A list with:
\enumerate{
\item The normalised X matrix in the first list element, and
\item An array of the corresponding dilution factors.
}
\itemize{
\item Following the example below will extract the results quickly and easily.
}
}
\description{
Histogram Matching (HM) is a reference based normalisation that aims to match the histogram created from a experimental spectra with that made from a reference spectra.
}
\details{
\subsection{How It Works:}{
\enumerate{
\item A more complex normalisation method compared to others in this package, HM creates a histogram of the intensities of an experimental spectrum.
\item This area of this histogram subtracted from that of a reference histogram, calculated from a median spectrum.
\item The experimental spectra is scaled by a factor \code{alpha} which alters the intensities and a new histogram is made and step 2. is repeated.
\item The aim is to minimise the difference between the reference histogram and the experiment histogram and the alpha value that acheives this minimum is the dilution coefficient.
\item This is iterated over for all experimetal spectra
}
}

\subsection{Advantages:}{
\itemize{
\item HM is not affected by peak shift like PQN (\code{\link[=pqNorm]{pqNorm()}}) is because it does not take into account the chemical shift of the intensity, on the value of the intensity.
\item HM also is not as impaacted by noise because noise is removed before histograms are produced (this is not touched on in the above section)
}
}

\subsection{Limitations:}{
\itemize{
\item Not all histograms are created equal, some will have more kurtosis or skewness which limits the degree to which differences can be minimised.
\item Studies have also found that HM does not handle noise well and does not recover signal like many other normalisation methods do. (see 'See also')
}
}
}
\examples{
# it is mandatory to fill args X, noi, and use_median for `hmNorm()` to work.
data(X, noi)
hmNorm(X, noi, use_median = TRUE, alpha = c(0.1, 2.5))
cat(dilf_hm)
}
\seealso{
\itemize{
\item The methods paper for HM can be found here: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}
\item The paper discussing HM limitations with noise can be found here: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}
}

Other {Reference-Based}: 
\code{\link{pqNorm}()},
\code{\link{xfNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
