% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqNorm.R
\name{pqNorm}
\alias{pqNorm}
\title{Probabilistic Quotient Normalisation}
\usage{
pqNorm(X)
}
\arguments{
\item{X}{A matrix containing the NMR spectra with rows contain all data points from one spectra and the columns containing the chemical shift variables.}
}
\value{
A list with the normalised X matrix and an array of the corresponding dilution factors.
}
\description{
Probabilistic Quotient Normalisation
}
\details{
The methods paper can be found here: https://doi.org/10.1021/ac051632c
}
\author{
\email{kylebario1@gmail.com}
}
