% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{standard preprocessing 1D}
\usage{
preprocessing(
  X,
  ppm,
  meta,
  flip = TRUE,
  cali = TRUE,
  baseline = TRUE,
  lowerCutoff,
  upperCutoff,
  waterLower,
  waterUpper,
  ureaLower,
  ureaUpper
)
}
\arguments{
\item{ppm}{an array of the chemical shift variables}
}
\description{
standard preprocessing 1D
}
\concept{preproc}
