% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roiNorm.R
\name{roiNorm}
\alias{roiNorm}
\title{Region of Interest Normalisation}
\usage{
roiNorm(X, sh = c(2.5, 2.75))
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{sh}{The numerical values defining the lower and upper regions of the Region of Interest. default = c(3,3.1).}
}
\value{
This function returns a list with:
\enumerate{
\item The normalised X matrix, and
\item A numerical array of the corresponding dilution factors.
}
\itemize{
\item Following the example below will extract the results quickly and easily.
}
}
\description{
This function normalises the spectra based on a specific area of the spectra. This is helpful when one region is constant across all spectra.
}
\details{
\subsection{How It Works:}{
\enumerate{
\item This function takes the values of the shift argument and finds the integral (or area) of the peak in that chemical shift region in an experimental spectra.
\item Each intensity in that experimental spectra is then divided by that integral to scale it.
This method works much like how \code{\link[=taNorm]{taNorm()}} and \code{\link[=creNorm]{creNorm()}} do, except instead of using an entire spectrum's integral or the creatinine peak specifically, you are able to use what ever ppm region you wish.
}
}

\subsection{Advantages:}{
\itemize{
\item Region of Interest Normalisation (ROI) gives the data analyst more control over what region will be used which can be advantageous.
}
}

\subsection{Limitations:}{
\itemize{
\item ROI is also subject to the flaws that impact creatinine normalisation (\code{\link[=creNorm]{creNorm()}}) in that, whatever factors that create variation in the region one chooses, will also impact the entire spectrum's state.
\item Reference-based normalisations may provide more robust normalisation.
}
}
}
\examples{
data(X)
roiNorm(X, sh = c(1.5,2))
cat(dilf_roi)
}
\seealso{
A description of Region of Interest Normalisation can be found in this paper: \url{http://dx.doi.org/10.1007/s11306-018-1400-6}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{qNorm2}()},
\code{\link{qNorm}()},
\code{\link{taNorm}()},
\code{\link{vecNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
