% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taNorm.R
\name{taNorm}
\alias{taNorm}
\title{Total Area Normalisation (TAN/CSN)}
\usage{
taNorm(X)
}
\arguments{
\item{X}{a numerical matrix with rows being the spectra and columns being the chemical shift variables}
}
\value{
a list of the normalised X matrix and a numerical array of the corresponding dilution factors
}
\description{
Total Area Normalisation (TAN/CSN)
}
\details{
also called Constant Sum normalisation https://doi.org/10.1021/ac051632c
}
\author{
kylebario1@gmail.com
}
