% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taNorm.R
\name{taNorm}
\alias{taNorm}
\title{Total Area Normalisation}
\usage{
taNorm(X)
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}
}
\value{
A list of the normalised X matrix and a numerical array of the corresponding dilution factors
}
\description{
\code{\link[=taNorm]{taNorm()}} finds the total integral of the spectra and divides each spectra by its integral.
}
\details{
Total area normalisation (TA) was once considered the gold standard in NMR spectra normalisation.
TA is useful for spectra that have similar peaks. If spectra differ a lot from each other - for instance, some have large peaks created by drug metabolites, etc. - then the total area will not be a suitable normalisation method.
TA is also referred to as constant sum or integral normalisation
}
\examples{
\dontrun{
    Xn <- taNorm(X)
    taDilf <- Xn[[2]]
    Xn <- Xn[[1]]}
}
\seealso{
\url{https://doi.org/10.1021/ac051632c}

Other {Attribute-Based}: 
\code{\link{creNorm}()},
\code{\link{roiNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Attribute-Based}}
