% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taNorm.R
\name{taNorm}
\alias{taNorm}
\title{Total Area Normalisation}
\usage{
taNorm(X)
}
\arguments{
\item{X}{spectral matrix with rows being the spectra and columns being the chemical shift variables}
}
\value{
list of normalised X matrix and an array of the corresponding dilution factors
}
\description{
Total Area Normalisation
}
\details{
also called Constant Sum normalisation
}
\author{
kylebario1@gmail.com
}
