% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfNorm.R
\name{xfNorm}
\alias{xfNorm}
\title{Normalisation based on an External Factor}
\usage{
xfNorm(X, xfactor)
}
\arguments{
\item{X}{a numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{xfactor}{a numerical array of values that correspond to the dilution coefficients of the spectra}
}
\value{
a list of the normalised X matrix and a numerical array of the corresponding dilution factors
}
\description{
Normalisation based on an External Factor
}
\details{
https://doi.org/10.1021/ac051632c
}
\examples{
  Xn <- xfNorm(X, osmo)
  xfDilf <- Xn[[2]]
  Xn <- Xn[[1]]
}
\author{
\email{kylebario1@gmail.com}
}
