% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfNorm.R
\name{xfNorm}
\alias{xfNorm}
\title{Normalisation based on an External Factor}
\usage{
xfNorm(X, xfactor)
}
\arguments{
\item{X}{A numerical matrix with rows being the spectra and columns being the chemical shift variables}

\item{xfactor}{A numerical array of values that correspond to the dilution coefficients of the spectra. Must be row-matched to the provided X matrix.}
}
\value{
This function returns a list with:
\enumerate{
\item The normalised X matrix in the first list element, and
\item A numerical array of the corresponding dilution factors in the second.
}
\itemize{
\item Following the example below will extract the results quickly and easily.
}
}
\description{
This function utilises an array of dilution coefficients acquired externally to normalise the spectra.
}
\details{
\subsection{How It Works:}{
\enumerate{
\item All the intensities of a spectrum are divided by its external dilution coefficient.
\item This is done to all spectra with their corresponding dilution factors
}
}

\subsection{Advantages:}{
\itemize{
\item Reliable: dilution estimations from external sources such as freezing-point depression osmometry are highly accurate, reducing the error in normalisation.
\item Computationally, this is not as intense as other normalisation methods like \code{\link[=pqNorm]{pqNorm()}} and \code{\link[=hmNorm]{hmNorm()}} and will take limited power to perform
}
}

\subsection{Limitations:}{
\itemize{
\item This method requires having a machine to provide the external measure of dilution such as a freezing-point depression osmometer.
\item Acquiring data on dilution with a manual freezing-point depression osmometer is a long and tedious process with each sample taking on average 5 minutes. Automatic osmometers are available but are expensive so this method is limited by logistics of actually acquiring the data.
}
}
}
\examples{
xf <- xfNorm(X, osmo)
Xn <- xf$Xn
dilf <- xf$dilf
}
\seealso{
\url{https://doi.org/10.1021/ac051632c}

Other {Reference-Based}: 
\code{\link{hmNorm}()},
\code{\link{pqNorm}()}
}
\author{
\email{kylebario1@gmail.com}
}
\concept{{Reference-Based}}
